Add: [기능] tool 파일로 분리
