+++ app/services/agents/restaurant_agent_service.py
@@ -1,250 +1,21 @@
 import traceback
-import os
 import json
-import asyncio
-import aiohttp
 from datetime import datetime
-from dotenv import load_dotenv
 from crewai import Agent, Task, Crew, LLM
-from crewai.tools import BaseTool
 from typing import List, Dict, Optional
 from fastapi import HTTPException
 from app.dtos.spot_models import spots_pydantic
+from dotenv import load_dotenv
+import os
+from app.services.agents.tools.restaurant_tool import (
+    GeocodingTool,
+    RestaurantBasicSearchTool,
+    NaverWebSearchTool,
+    NaverImageSearchTool,
+)
 
-# 환경 변수 로드
 load_dotenv()
 OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
-GOOGLE_MAP_API_KEY = os.getenv("GOOGLE_MAP_API_KEY")
-AGENT_NAVER_CLIENT_ID = os.getenv("AGENT_NAVER_CLIENT_ID")
-AGENT_NAVER_CLIENT_SECRET = os.getenv("AGENT_NAVER_CLIENT_SECRET")
-
-llm = LLM(model="gpt-4o", temperature=0, api_key=OPENAI_API_KEY)
-
-# ------------------------- 비동기 Tool 클래스 -------------------------
-# 1. Google Geocoding API를 사용하여 좌표를 조회하는 Tool
-class GeocodingTool(BaseTool):
-    name: str = "GeocodingTool"
-    description: str = (
-        "Google Geocoding API를 사용하여 주어진 위치의 위도와 경도를 조회합니다. "
-        "입력된 location 값은 변경 없이 그대로 반환합니다."
-    )
-
-    async def _arun(self, location: str) -> Dict:
-        url = "https://maps.googleapis.com/maps/api/geocode/json"
-        params = {"address": location, "key": GOOGLE_MAP_API_KEY}
-        try:
-            async with aiohttp.ClientSession() as session:
-                async with session.get(url, params=params) as response:
-                    data = await response.json()
-                    if data.get("results"):
-                        loc = data["results"][0]["geometry"]["location"]
-                        coordinates = f"{loc['lat']},{loc['lng']}"
-                    else:
-                        coordinates = ""
-        except Exception as e:
-            coordinates = f"[GeocodingTool] Error: {str(e)}"
-        return {"location": location, "coordinates": coordinates}
-
-    def _run(self, location: str) -> Dict:
-        return asyncio.run(self._arun(location))
-
-
-# 2. Google Places API를 사용해 맛집 기본 정보를 조회하는 Tool
-class RestaurantBasicSearchTool(BaseTool):
-    name: str = "RestaurantBasicSearchTool"
-    description: str = (
-        "주어진 좌표와 location 정보를 기반으로 구글맵에서 식당의 title, rating, reviews를 검색합니다."
-    )
-
-    async def get_place_details(
-        self, session: aiohttp.ClientSession, place_id: str
-    ) -> Dict:
-        url = "https://maps.googleapis.com/maps/api/place/details/json"
-        params = {
-            "place_id": place_id,
-            "fields": "name,rating,user_ratings_total,geometry",
-            "language": "ko",
-            "key": GOOGLE_MAP_API_KEY,
-        }
-        try:
-            async with session.get(url, params=params) as response:
-                data = await response.json()
-                result = data.get("result", {})
-                return {
-                    "title": result.get("name"),
-                    "rating": result.get("rating", 0),
-                    "reviews": result.get("user_ratings_total", 0),
-                    "latitude": result["geometry"]["location"]["lat"],
-                    "longitude": result["geometry"]["location"]["lng"],
-                }
-        except Exception as e:
-            print(f"[RestaurantBasicSearchTool] Details Error: {e}")
-            return None
-
-    async def _arun(self, location: str, coordinates: str) -> List[Dict]:
-        url = "https://maps.googleapis.com/maps/api/place/textsearch/json"
-        all_candidates = []
-        lat, lng = coordinates.split(",")
-        params = {
-            "query": f"{location} 맛집",
-            "language": "ko",
-            "type": "restaurant",
-            "location": f"{lat},{lng}",
-            "radius": "5000",
-            "key": GOOGLE_MAP_API_KEY,
-        }
-        try:
-            async with aiohttp.ClientSession() as session:
-                # 첫 번째 요청
-                async with session.get(url, params=params) as response:
-                    data = await response.json()
-                    print(f"첫 요청 결과 수: {len(data.get('results', []))}")
-                    for place in data.get("results", []):
-                        place_id = place.get("place_id")
-                        if place_id:
-                            details = await self.get_place_details(session, place_id)
-                            if (
-                                details
-                                and details["rating"] >= 4.0
-                                and details["reviews"] >= 500
-                            ):
-                                all_candidates.append(details)
-                    next_page_token = data.get("next_page_token")
-
-                # 추가 요청: 후보 수가 10개 미만이면 추가로 요청
-                while next_page_token and len(all_candidates) < 15:
-                    try:
-                        await asyncio.sleep(3)  # next_page_token 유효 대기
-                        params["pagetoken"] = next_page_token
-                        async with session.get(url, params=params) as response:
-                            data = await response.json()
-                            new_results = data.get("results", [])
-                            print(f"추가 요청 결과 수: {len(new_results)}")
-                            for place in new_results:
-                                # 상한선(예: 40개)은 기존 코드와 동일하게 유지
-                                if len(all_candidates) >= 40:
-                                    break
-                                place_id = place.get("place_id")
-                                if place_id:
-                                    details = await self.get_place_details(
-                                        session, place_id
-                                    )
-                                    if (
-                                        details
-                                        and details["rating"] >= 4.0
-                                        and details["reviews"] >= 500
-                                    ):
-                                        all_candidates.append(details)
-                            next_page_token = data.get("next_page_token")
-                    except Exception as e:
-                        print(f"추가 페이지 요청 오류: {e}")
-                        break
-
-                print(f"최종 수집된 맛집 수: {len(all_candidates)}")
-        except Exception as e:
-            print(f"[RestaurantBasicSearchTool] Search Error: {e}")
-        return all_candidates
-
-    def _run(self, location: str, coordinates: str) -> List[Dict]:
-        return asyncio.run(self._arun(location, coordinates))
-
-
-# 3. 네이버 웹 검색 API를 사용해 식당의 세부 정보를 조회하는 Tool
-class NaverWebSearchTool(BaseTool):
-    name: str = "NaverWebSearch"
-    description: str = "네이버 웹 검색 API를 사용해 식당의 상세 정보를 검색합니다."
-
-    async def fetch(self, session: aiohttp.ClientSession, query: str):
-        url = "https://openapi.naver.com/v1/search/webkr.json"
-        headers = {
-            "X-Naver-Client-Id": AGENT_NAVER_CLIENT_ID,
-            "X-Naver-Client-Secret": AGENT_NAVER_CLIENT_SECRET,
-        }
-        print(f"[네이버 세부정부 검색어]: {query}")
-        params = {
-            "query": query,
-            "display": 3,
-            "start": 1,
-            "sort": "sim",
-        }
-        try:
-            async with session.get(url, headers=headers, params=params) as response:
-                data = await response.json()
-                items = data.get("items", [])
-                if not items:
-                    return {"description": "정보를 찾을 수 없습니다.", "url": ""}
-                descriptions = []
-                for item in items:
-                    desc = item.get("description", "").strip()
-                    if desc and len(desc) > 30:
-                        descriptions.append(desc)
-                combined_description = " ".join(descriptions)
-                return {
-                    "description": (
-                        combined_description[:200]
-                        if len(combined_description) > 200
-                        else combined_description
-                    ),
-                    "url": items[0].get("link", "") if items else "",
-                }
-        except Exception as e:
-            print(f"네이버 웹 검색 오류: {str(e)}")
-            return {"description": "정보 없음", "url": ""}
-
-    async def _arun(self, restaurant_list: List[str]) -> Dict[str, Dict[str, str]]:
-        results = {}
-        async with aiohttp.ClientSession() as session:
-            for restaurant in restaurant_list:
-                results[restaurant] = await self.fetch(session, restaurant)
-        return results
-
-    def _run(self, restaurant_list: List[str]) -> Dict[str, Dict[str, str]]:
-        return asyncio.run(self._arun(restaurant_list))
-
-
-# 4. 네이버 이미지 검색 API를 사용해 식당의 대표 이미지를 조회하는 Tool
-class NaverImageSearchTool(BaseTool):
-    name: str = "NaverImageSearch"
-    description: str = (
-        "네이버 이미지 검색 API를 사용해 식당의 대표 이미지를 검색합니다."
-    )
-
-    async def fetch(self, session: aiohttp.ClientSession, query: str):
-        url = "https://openapi.naver.com/v1/search/image"
-        headers = {
-            "X-Naver-Client-Id": AGENT_NAVER_CLIENT_ID,
-            "X-Naver-Client-Secret": AGENT_NAVER_CLIENT_SECRET,
-        }
-        print(f"[네이버 이미지 검색어]: {query}")
-        params = {
-            "query": query,
-            "display": 5,
-            "sort": "sim",
-            "filter": "large",
-        }
-        try:
-            async with session.get(url, headers=headers, params=params) as response:
-                data = await response.json()
-                items = data.get("items", [])
-                if not items:
-                    return "https://via.placeholder.com/300x200?text=No+Image"
-                return items[0].get(
-                    "link", "https://via.placeholder.com/300x200?text=No+Image"
-                )
-        except Exception as e:
-            print(f"네이버 이미지 검색 오류: {str(e)}")
-            return "https://via.placeholder.com/300x200?text=Error"
-
-    async def _arun(self, restaurant_list: List[str]) -> Dict[str, str]:
-        results = {}
-        async with aiohttp.ClientSession() as session:
-            for restaurant in restaurant_list:
-                results[restaurant] = await self.fetch(session, restaurant)
-        return results
-
-    def _run(self, restaurant_list: List[str]) -> Dict[str, str]:
-        return asyncio.run(self._arun(restaurant_list))
-
 
 # ------------------------- Restaurant Agent Service -------------------------
 class RestaurantAgentService:
@@ -262,12 +33,10 @@ class RestaurantAgentService:
         """서비스 초기화"""
         self.llm = LLM(model="gpt-4o", temperature=0, api_key=OPENAI_API_KEY)
         # Tools 초기화
-        self.tools = {
-            "geocoding": GeocodingTool(),
-            "restaurant_search": RestaurantBasicSearchTool(),
-            "web_search": NaverWebSearchTool(),
-            "image_search": NaverImageSearchTool(),
-        }
+        self.geocoding_tool = GeocodingTool()
+        self.restaurant_search_tool = RestaurantBasicSearchTool()
+        self.web_search_tool = NaverWebSearchTool()
+        self.image_search_tool = NaverImageSearchTool()
         self.agents = self._create_agents()
 
     def _process_input(
@@ -309,7 +78,7 @@ class RestaurantAgentService:
                 role="좌표 조회 전문가",
                 goal="사용자가 입력한 location(예: '부산광역시')의 위도와 경도를 조회하며, location 값은 그대로 유지한다.",
                 backstory="나는 위치 데이터 전문가로, 입력된 location 값을 변경하지 않고 Google Geocoding API를 통해 좌표를 조회한다.",
-                tools=[self.tools["geocoding"]],
+                tools=[self.geocoding_tool],
                 llm=self.llm,
                 verbose=True,
                 async_execution=True,
@@ -318,7 +87,7 @@ class RestaurantAgentService:
                 role="맛집 기본 조회 전문가",
                 goal="좌표 정보를 활용하여 식당의 기본 정보를 조회한다.",
                 backstory="나는 맛집 데이터 분석 전문가로, Google Maps API를 사용하여 특정 위치의 식당 정보를 최대 40개까지 조회한다.",
-                tools=[self.tools["restaurant_search"]],
+                tools=[self.restaurant_search_tool],
                 llm=self.llm,
                 verbose=True,
                 async_execution=True,
@@ -327,7 +96,7 @@ class RestaurantAgentService:
                 role="네이버 웹 검색 에이전트",
                 goal="네이버 웹 검색 API를 사용해 식당의 텍스트 기반 세부 정보를 조회한다.",
                 backstory="나는 네이버 웹 검색 전문가로, 식당의 상세 텍스트 정보를 제공합니다.",
-                tools=[self.tools["web_search"]],
+                tools=[self.web_search_tool],
                 llm=self.llm,
                 verbose=True,
                 async_execution=True,
@@ -336,7 +105,7 @@ class RestaurantAgentService:
                 role="네이버 이미지 검색 에이전트",
                 goal="네이버 이미지 검색 API를 사용해 식당의 이미지 URL을 조회한다.",
                 backstory="나는 네이버 이미지 검색 전문가로, 식당의 정확한 이미지를 제공합니다.",
-                tools=[self.tools["image_search"]],
+                tools=[self.image_search_tool],
                 llm=self.llm,
                 verbose=True,
                 async_execution=True,
@@ -345,7 +114,7 @@ class RestaurantAgentService:
                 role="최종 추천 에이전트",
                 goal="필터링된 맛집 후보와 네이버 텍스트 기반 세부 정보를, 여행 계획을 고려하여 최종 맛집 추천 리스트를 생성한다.",
                 backstory="나는 데이터 구조화 전문가로, 후보 식당의 기본 정보, 네이버에서 수집한 텍스트 세부 정보와 여행 계획 정보를 종합하여 최종 추천 리스트를 구성한다.",
-                tools=[self.tools["web_search"]],
+                tools=[self.web_search_tool],
                 llm=self.llm,
                 verbose=True,
                 async_execution=True,
